+++ src/utils/random-number-creator.go
@@ -2,9 +2,9 @@ package custom_utils
 
 import "math/rand"
 
-func CreateRandomNumber(min int, max int) int {
+func CreateRandomNumber(min int64, max int64) int64 {
 	if min == max {
 		return min
 	}
-	return rand.Intn(max-min) + min
+	return int64(rand.Intn(int(max-min))) + min
 }
